api/requirements.txt
fastapi
uvicorn
pydantic
pymongo
boto3
requests
python-multipart
python-dotenv
pydantic-settings
python-jose
passlib
deepgram-sdk